requirements.txt
# Core dependencies
gradio>=4.0.0
pandas>=1.5.0
numpy>=1.24.0

# Vision Language Model dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.20.0
Pillow>=9.5.0

# Additional ML dependencies
sentence-transformers>=2.2.0
datasets>=2.14.0

# File processing
requests>=2.31.0
opencv-python>=4.8.0
python-magic>=0.4.27

# Output generation
fpdf2>=2.7.0
python-docx>=0.8.11
markdown>=3.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pathlib
typing_extensions>=4.5.0